libteam: add library to control team devices
The libteam package provides LACP support in Linux. A patch for netifd
has been submitted that provides configuration support for team devices
using this package. The provided daemon manages a kernel team device
that provides software LAG or hardware offloaded LAG when the switch
driver supports that.

The new package requires a (submitted) patch to libnl base package to
provide the libnl-cli library.

Signed-off-by: Pavel Šimerda <code@simerda.eu>